Update and rename Askisi1b_icsd15111.s to Example-1/Example1b_icsd15111.s
Example 1B
Questions:
2. Write a program that asks the user to enter two integers numbers, displaying appropriate messages to the console QtSpim.
Then will performs the act of removing the second number from the first and displays the effect, accompanied by an appropriate message (eg "The result of the subtraction is"), in the console. To store the two numbers and the result of subtraction to total use two registers of MIPS, the $ t0 and $ t1.